requirements.txt
# FastAPI & ASGI Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database & ORM
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.5

# Google Cloud Platform
google-cloud-storage==2.18.2
google-cloud-pubsub==2.24.0
google-cloud-sql-connector==1.12.0
google-genai==1.5.0

# AI & Machine Learning
numpy==2.1.0
pillow==10.4.0

# Utilities
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
